requirements.txt
asttokens==2.2.1
backcall==0.2.0
bbox-visualizer==0.1.0
comm==0.1.3
contourpy==1.1.0
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
fonttools==4.41.0
ipykernel==6.24.0
ipython==8.14.0
ipywidgets==8.0.7
jedi==0.18.2
joblib==1.3.1
jupyter-bbox-widget==0.5.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyterlab-widgets==3.0.8
kiwisolver==1.4.4
matplotlib==3.7.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.25.1
opencv-python==4.8.0.74
packaging==23.1
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.8.1
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
pylabel==0.1.51
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.0
scikit-learn==1.3.0
scipy==1.11.1
six==1.16.0
stack-data==0.6.2
threadpoolctl==3.2.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
tzdata==2023.3
wcwidth==0.2.6
widgetsnbextension==4.0.8
